Vendor catalog.
Ditch the submodule.  The long term idea isn't to use submodules; so
let's stop *now*.  It's a really weird thing to have the first demo
in the docs demonstrate a usage pattern that's Not Important.

Vendor the busybash image lineage info.  The name used by the module
file also changes; this is the name presently in the upstream catalog
at https://github.com/polydawn/catalog (the submodule we just removed).

This also has the 'lineage.tl' filename instead of the older
'catalog.tl' filename, so this commit brings us up to date to work with
Reach (nee Stellar; I have to rename this repo momentarily as well) and
the Timeless Stack v0.300 release.

Signed-off-by: Eric Myhre <hash@exultant.us>